Tools: Tune: TDFB: Add generate of few needed blobs
This patch adds generate of two mic line array with 74 mm
spacing to example_line_array.m.

The generate of passthrough blobs is updated in script
example_pass_config.m. Without this the current beamformer rejects
the blobs as too old version. There needs to be at least one beam
angle defined. A blob with 4 mic to 2 channels passthrough is added.
It selects channels 1 4 into the stereo stream. Those would be the
leftmost and rightmost microphone in devices with 4 microphones.

A new script example_two_beams_default.m is added. It creates a
stereo two beams beamformer for 2 and 4 mic systems. The stereo
angle is very narrow for user focused audio.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>